backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
httpx==0.26.0
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.17
pinecone-client==3.0.0
transformers==4.37.2
torch==2.2.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
yfinance==0.2.36
beautifulsoup4==4.12.3
textblob==0.17.1
